refactor: add and use new generic postgres driver
