+++ database_sql_public_ip_test.go
@@ -19,7 +19,7 @@ import (
 	"fmt"
 
 	"cloud.google.com/go/alloydbconn"
-	"cloud.google.com/go/alloydbconn/driver/pgxv5"
+	"cloud.google.com/go/alloydbconn/driver/postgres"
 )
 
 // connectDatabaseSQLWithPublicIP establishes a connection to your database
@@ -44,7 +44,7 @@ func connectDatabaseSQLWithPublicIP(
 ) (*sql.DB, func() error, error) {
 	// First, register the AlloyDB driver. Note, the driver's name is arbitrary
 	// and must only match what you use below in sql.Open. Also,
-	// pgxv5.RegisterDriver accepts options to configure credentials, timeouts,
+	// postgres.RegisterDriver accepts options to configure credentials, timeouts,
 	// etc.
 	//
 	// For details, see:
@@ -54,7 +54,7 @@ func connectDatabaseSQLWithPublicIP(
 	// goroutines. Call it when you're done with your database connection to
 	// avoid a goroutine leak.
 	opts = append(opts, alloydbconn.WithDefaultDialOptions(alloydbconn.WithPublicIP()))
-	cleanup, err := pgxv5.RegisterDriver("alloydb-public", opts...)
+	cleanup, err := postgres.RegisterDriver("alloydb-public", opts...)
 	if err != nil {
 		return nil, cleanup, err
 	}
