+++ driver/pgxv5/postgres.go
@@ -14,18 +14,13 @@
 
 // Package pgxv5 provides an AlloyDB driver that uses pgx v5 and works with the
 // database/sql package.
+//
+// Deprecated: Use cloud.google.com/go/alloydbconn/driver/postgres instead.
 package pgxv5
 
 import (
-	"context"
-	"database/sql"
-	"database/sql/driver"
-	"net"
-	"sync"
-
 	"cloud.google.com/go/alloydbconn"
-	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/stdlib"
+	"cloud.google.com/go/alloydbconn/driver/postgres"
 )
 
 // RegisterDriver registers a Postgres driver that uses the alloydbconn.Dialer
@@ -34,61 +29,8 @@ import (
 // differently configured Dialers. The driver uses pgx/v5 internally.
 // RegisterDriver returns a cleanup function that should be called one the
 // database connection is no longer needed.
-func RegisterDriver(name string, opts ...alloydbconn.Option) (func() error, error) {
-	d, err := alloydbconn.NewDialer(context.Background(), opts...)
-	if err != nil {
-		return func() error { return nil }, err
-	}
-	sql.Register(name, &pgDriver{
-		d:      d,
-		dbURIs: make(map[string]string),
-	})
-	return func() error { return d.Close() }, nil
-}
-
-type pgDriver struct {
-	d  *alloydbconn.Dialer
-	mu sync.RWMutex
-	// dbURIs is a map of DSN to DB URI for registered connection names.
-	dbURIs map[string]string
-}
-
-// Open accepts a keyword/value formatted connection string and returns a
-// connection to the database using alloydbconn.Dialer. The AlloyDB instance URI
-// should be specified in the host field. For example:
 //
-// "host=projects/<PROJECT>/locations/<REGION>/clusters/<CLUSTER>/instances/<INSTANCE> user=myuser password=mypass"
-func (p *pgDriver) Open(name string) (driver.Conn, error) {
-	dbURI, err := p.dbURI(name)
-	if err != nil {
-		return nil, err
-	}
-	return stdlib.GetDefaultDriver().Open(dbURI)
-
-}
-
-// dbURI registers a driver using the provided DSN. If the name has already
-// been registered, dbURI returns the existing registration.
-func (p *pgDriver) dbURI(name string) (string, error) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	dbURI, ok := p.dbURIs[name]
-	if ok {
-		return dbURI, nil
-	}
-
-	config, err := pgx.ParseConfig(name)
-	if err != nil {
-		return "", err
-	}
-	instConnName := config.Config.Host // Extract instance connection name
-	config.Config.Host = "localhost"   // Replace it with a default value
-	config.DialFunc = func(ctx context.Context, _, _ string) (net.Conn, error) {
-		return p.d.Dial(ctx, instConnName)
-	}
-
-	dbURI = stdlib.RegisterConnConfig(config)
-	p.dbURIs[name] = dbURI
-
-	return dbURI, nil
+// Deprecated: Use postgres.RegisterDriver instead.
+func RegisterDriver(name string, opts ...alloydbconn.Option) (func() error, error) {
+	return postgres.RegisterDriver(name, opts...)
 }
